Build embedded documentation as lib
Thinking about CI/CD, the current setup is not yet perfect.
`app`, `docs` and `server` need to be built separately with some kind of script.
`app` and `docs` need to be built before `server`, such that the embedding works correctly.
Ideally, we want these dependencies to be handled by our build tools.
Not only is it less error prone, but our CI/CD pipeline will also be simplified.
This commit wraps the embedded documentation in a Rust crate, solving half of the problem.
The `app` is more difficult to do, because `trunk` calls `cargo` under the hood,
which causes a deadlock over the build directory if `trunk` is called from `build.rs`.