Use OUT_DIR idiomatically in build script
Previously, the build.rs of the documentation would modify the working
directory of the project. This is not good, build scripts should
only write to the directory in the OUT_DIR variable.

There appears to be a little extra code-gen required to be able
to point rust-embed at the OUT_DIR.