Statically compile for the x86_64-musl target. (!23)
This enables the use of a scratch container, bringing the final size
down to around 11 MB. Note that native dependencies like
sqlite can lead to problems (solvable but annoying). The same holds true
when compiling for different architectures than x86_64.

Instead of the sed command one could directly add the musl target to the
rust-toolchain.toml file. But as it is only needed in the container
build so far, this seems to be the right way for now.