Get measurements from microcontroller (!37)
On one hand this enables the microcontroller to send measurements to an MQTT broker, on the other hand the server component subscribes to new measurements from the same broker.

Design decisions for the microcontroller part are described in the documentation (mdbook). More information about configuring the microcontroller is in the `controller` subproject readme. A minimal example setup is described in the documentation as well (chapter 6).